requirements.txt
asgiref==3.6.0
dj-database-url==1.2.0
Django==4.1.6
django-crontab==0.7.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
Pillow==9.4.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
python-dotenv==0.21.1
pytz==2022.7.1
sqlparse==0.4.3
whitenoise==6.3.0
